[Chore] #14 회원가입 userController로 위치 변경
